gitter: Fix occupants for GitterRoom
The limit parameter was not used which defaults to 30
and returns maximum 100 users at a time. Hence, it fails
to return all the members in the room.
In order to tackle this, we need to iterate by skipping
100 members with every API request. Also users count
needs to be updated everytime list of occupants is
requested as new users may join a room in realtime.

Closes https://github.com/errbotio/err-backend-gitter/issues/38